pyproject.toml
[tool.poetry]
name = "code_executor"
version = "0.1.0"
description = ""
authors = ["timur <timurza@icloud.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.0"
uvicorn = "^0.23.2"
pydantic-settings = "^2.0.3"
python-jose = "^3.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
